Let DAGContext accepts Store instead of Snapshot
