+++ src/coprocessor/dag/pipeline.rs
@@ -1,0 +1,106 @@
+// Copyright 2019 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::sync::Arc;
+
+use kvproto::coprocessor::KeyRange;
+use tipb::executor::{self, ExecType};
+
+use storage::Store;
+
+use super::executor::{
+    Executor, HashAggExecutor, IndexScanExecutor, LimitExecutor, SelectionExecutor,
+    StreamAggExecutor, TableScanExecutor, TopNExecutor,
+};
+use coprocessor::dag::expr::EvalConfig;
+use coprocessor::*;
+
+pub struct ExecutorPipelineBuilder;
+
+impl ExecutorPipelineBuilder {
+    pub fn build_normal<S: Store + 'static>(
+        execs: Vec<executor::Executor>,
+        store: S,
+        ranges: Vec<KeyRange>,
+        ctx: Arc<EvalConfig>,
+        collect: bool,
+    ) -> Result<Box<Executor + Send>> {
+        let mut execs = execs.into_iter();
+        let first = execs
+            .next()
+            .ok_or_else(|| Error::Other(box_err!("has no executor")))?;
+        let mut src = Self::build_normal_first_executor(first, store, ranges, collect)?;
+        for mut exec in execs {
+            let curr: Box<Executor + Send> = match exec.get_tp() {
+                ExecType::TypeTableScan | ExecType::TypeIndexScan => {
+                    return Err(box_err!("got too much *scan exec, should be only one"))
+                }
+                ExecType::TypeSelection => Box::new(SelectionExecutor::new(
+                    exec.take_selection(),
+                    Arc::clone(&ctx),
+                    src,
+                )?),
+                ExecType::TypeAggregation => Box::new(HashAggExecutor::new(
+                    exec.take_aggregation(),
+                    Arc::clone(&ctx),
+                    src,
+                )?),
+                ExecType::TypeStreamAgg => Box::new(StreamAggExecutor::new(
+                    Arc::clone(&ctx),
+                    src,
+                    exec.take_aggregation(),
+                )?),
+                ExecType::TypeTopN => {
+                    Box::new(TopNExecutor::new(exec.take_topN(), Arc::clone(&ctx), src)?)
+                }
+                ExecType::TypeLimit => Box::new(LimitExecutor::new(exec.take_limit(), src)),
+            };
+            src = curr;
+        }
+        Ok(src)
+    }
+
+    fn build_normal_first_executor<S: Store + 'static>(
+        mut first: executor::Executor,
+        store: S,
+        ranges: Vec<KeyRange>,
+        collect: bool,
+    ) -> Result<Box<Executor + Send>> {
+        match first.get_tp() {
+            ExecType::TypeTableScan => {
+                let ex = Box::new(TableScanExecutor::new(
+                    first.take_tbl_scan(),
+                    ranges,
+                    store,
+                    collect,
+                )?);
+                Ok(ex)
+            }
+            ExecType::TypeIndexScan => {
+                let unique = first.get_idx_scan().get_unique();
+                let ex = Box::new(IndexScanExecutor::new(
+                    first.take_idx_scan(),
+                    ranges,
+                    store,
+                    unique,
+                    collect,
+                )?);
+                Ok(ex)
+            }
+            _ => Err(box_err!(
+                "first exec type should be *Scan, but get {:?}",
+                first.get_tp()
+            )),
+        }
+    }
+}
