+++ src/coprocessor/endpoint.rs
@@ -25,7 +25,7 @@ use tipb::select::DAGRequest;
 
 use server::readpool::{self, ReadPool};
 use server::Config;
-use storage::{self, Engine};
+use storage::{self, Engine, SnapshotStore};
 use util::Either;
 
 use coprocessor::dag::executor::ExecutorMetrics;
@@ -125,10 +125,15 @@ impl<E: Engine> Endpoint<E> {
                     Some(dag.get_start_ts()),
                 );
                 let batch_row_limit = self.get_batch_row_limit(is_streaming);
-                builder = box move |snap, req_ctx: &_| {
-                    // See rust-lang#41078 to know why we have `: &_` here.
-                    dag::DAGContext::new(dag, ranges, snap, req_ctx, batch_row_limit)
-                        .map(|h| h.into_boxed())
+                builder = box move |snap, req_ctx: &ReqContext| {
+                    // TODO: Remove explicit type once rust-lang#41078 is resolved
+                    let store = SnapshotStore::new(
+                        snap,
+                        dag.get_start_ts(),
+                        req_ctx.context.get_isolation_level(),
+                        !req_ctx.context.get_not_fill_cache(),
+                    );
+                    dag::DAGContext::build(dag, ranges, store, req_ctx.deadline, batch_row_limit)
                 };
             }
             REQ_TYPE_ANALYZE => {
@@ -145,6 +150,7 @@ impl<E: Engine> Endpoint<E> {
                     Some(analyze.get_start_ts()),
                 );
                 builder = box move |snap, req_ctx: &_| {
+                    // TODO: Remove explicit type once rust-lang#41078 is resolved
                     statistics::analyze::AnalyzeContext::new(analyze, ranges, snap, req_ctx)
                         .map(|h| h.into_boxed())
                 };
@@ -163,6 +169,7 @@ impl<E: Engine> Endpoint<E> {
                     Some(checksum.get_start_ts()),
                 );
                 builder = box move |snap, req_ctx: &_| {
+                    // TODO: Remove explicit type once rust-lang#41078 is resolved
                     checksum::ChecksumContext::new(checksum, ranges, snap, req_ctx)
                         .map(|h| h.into_boxed())
                 };
