+++ src/coprocessor/mod.rs
@@ -59,7 +59,7 @@ const SINGLE_GROUP: &[u8] = b"SingleGroup";
 type HandlerStreamStepResult = Result<(Option<coppb::Response>, bool)>;
 
 /// An interface for all kind of Coprocessor request handlers.
-trait RequestHandler: Send {
+pub trait RequestHandler: Send {
     /// Processes current request and produces a response.
     fn handle_request(&mut self) -> Result<coppb::Response> {
         panic!("unary request is not supported for this handler");
@@ -75,7 +75,7 @@ trait RequestHandler: Send {
         // Do nothing by default
     }
 
-    fn into_boxed(self) -> Box<RequestHandler + Send>
+    fn into_boxed(self) -> Box<RequestHandler>
     where
         Self: 'static + Sized,
     {
@@ -122,7 +122,7 @@ impl Deadline {
 /// Denotes for a function that builds a `RequestHandler`.
 /// Due to rust-lang#23856, we have to make it a type alias of `Box<..>`.
 type RequestHandlerBuilder<Snap> =
-    Box<dyn for<'a> FnBox(Snap, &'a ReqContext) -> Result<Box<dyn RequestHandler + Send>> + Send>;
+    Box<dyn for<'a> FnBox(Snap, &'a ReqContext) -> Result<Box<dyn RequestHandler>> + Send>;
 
 /// Encapsulate the `kvrpcpb::Context` to provide some extra properties.
 #[derive(Debug, Clone)]
